Serialization of graphs working now, but not tested fully
--HG--
extra : convert_revision : svn%3Ad1681fee-8e23-fc4c-b001-5eb8f207e95b/trunk%4094

+++ _Program/Serialization/Xml/IXmlSerializationInfo.cs
@@ -20,8 +20,8 @@ namespace Altaxo.Serialization.Xml
 		int GetInt32(string name);
 		void AddValue(string name, int val);
 
-		double GetSingle();
-		double GetSingle(string name);
+		float GetSingle();
+		float GetSingle(string name);
 		void AddValue(string name, float val);
 
 		double GetDouble();
